liquidsoap/python/requirements.txt
mutagen
pillow
